reorganize contents
