+++ dango/orderbook/src/core/order_matching.rs
@@ -1,0 +1,94 @@
+use {
+    crate::Order,
+    dango_types::orderbook::OrderId,
+    grug::{Number, NumberConst, StdResult, Udec128, Uint128},
+};
+
+pub struct MatchingOutcome {
+    /// The range of prices that achieve the biggest trading volume.
+    /// `None` if no match is found.
+    ///
+    /// All prices in this range achieve the same volume. It's up to the caller
+    /// to decide which price to use: the lowest, the highest, or the midpoint.
+    pub range: Option<(Udec128, Udec128)>,
+    /// The amount of trading volume, measured as the amount of the base asset.
+    pub volume: Uint128,
+    /// The BUY orders that have found a match.
+    pub bids: Vec<((Udec128, OrderId), Order)>,
+    /// The SELL orders that have found a match.
+    pub asks: Vec<((Udec128, OrderId), Order)>,
+}
+
+/// Given the standing BUY and SELL orders in the book, find range of prices
+/// that maximizes the trading volume.
+///
+/// ## Inputs:
+///
+/// - `bid_iter`: An iterator over the BUY orders in the book. This should
+///   follow the **price-time priority**, meaning it should return the order
+///   with the best price (in the case of BUY orders, the highest price) first;
+///   for orders the same price, the oldest one first.
+/// - `ask_iter`: An iterator over the SELL orders in the book that similarly
+///   follows the price-time priority.
+pub fn match_orders<B, A>(mut bid_iter: B, mut ask_iter: A) -> StdResult<MatchingOutcome>
+where
+    B: Iterator<Item = StdResult<((Udec128, OrderId), Order)>>,
+    A: Iterator<Item = StdResult<((Udec128, OrderId), Order)>>,
+{
+    let mut bid = bid_iter.next().transpose()?;
+    let mut bids = Vec::new();
+    let mut bid_is_new = true;
+    let mut bid_volume = Uint128::ZERO;
+    let mut ask = ask_iter.next().transpose()?;
+    let mut asks = Vec::new();
+    let mut ask_is_new = true;
+    let mut ask_volume = Uint128::ZERO;
+    let mut range = None;
+
+    loop {
+        let Some(((bid_price, bid_order_id), bid_order)) = bid else {
+            break;
+        };
+
+        let Some(((ask_price, ask_order_id), ask_order)) = ask else {
+            break;
+        };
+
+        if bid_price < ask_price {
+            break;
+        }
+
+        range = Some((ask_price, bid_price));
+
+        if bid_is_new {
+            bids.push(((bid_price, bid_order_id), bid_order));
+            bid_volume.checked_add_assign(bid_order.remaining)?;
+        }
+
+        if ask_is_new {
+            asks.push(((ask_price, ask_order_id), ask_order));
+            ask_volume.checked_add_assign(ask_order.remaining)?;
+        }
+
+        if bid_volume <= ask_volume {
+            bid = bid_iter.next().transpose()?;
+            bid_is_new = true;
+        } else {
+            bid_is_new = false;
+        }
+
+        if ask_volume <= bid_volume {
+            ask = ask_iter.next().transpose()?;
+            ask_is_new = true;
+        } else {
+            ask_is_new = false;
+        }
+    }
+
+    Ok(MatchingOutcome {
+        range,
+        volume: bid_volume.min(ask_volume),
+        bids,
+        asks,
+    })
+}
