+++ dango/orderbook/src/lib.rs
@@ -1,6 +1,6 @@
-mod clearing;
+mod core;
 mod execute;
 mod query;
 mod state;
 
-pub use {clearing::*, execute::*, query::*, state::*};
+pub use {core::*, execute::*, query::*, state::*};
